begin the rebuild cosn
build the shell

get through the shell to original cosn

add the shell of cosn to use ofs to support merge bucket

change the read and write thread class loader solve the stack issue

change the version begin to release the unite

update chdfs network version

add the interface to judge whehter is merge bucket

add interface to judge whether cos process

add truncate proxy

close ranger client when close file system

add release interface to merge bucket

rename merge bucket to posix bucket
