+++ src/main/java/org/apache/hadoop/fs/CosFileSystem.java
@@ -1,16 +1,14 @@
 package org.apache.hadoop.fs;
 
 import com.google.common.base.Preconditions;
-import com.google.common.util.concurrent.ThreadFactoryBuilder;
+import com.qcloud.chdfs.fs.CHDFSHadoopFileSystemAdapter;
 import com.qcloud.cos.model.HeadBucketResult;
 import com.qcloud.chdfs.permission.RangerAccessType;
-import com.qcloud.cos.utils.StringUtils;
-import org.apache.hadoop.HadoopIllegalArgumentException;
 import org.apache.hadoop.classification.InterfaceAudience;
 import org.apache.hadoop.classification.InterfaceStability;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.auth.RangerCredentialsProvider;
-import org.apache.hadoop.fs.cosn.*;
+import org.apache.hadoop.fs.cosn.Constants;
 import org.apache.hadoop.fs.cosn.ranger.client.RangerQcloudObjectStorageClient;
 import org.apache.hadoop.fs.cosn.ranger.security.authorization.AccessType;
 import org.apache.hadoop.fs.cosn.ranger.security.authorization.PermissionRequest;
@@ -43,6 +41,10 @@ import java.util.concurrent.*;
  * . Unlike
  * {@link CosFileSystem} this implementation stores files on COS in their
  * native form so they can be read by other cos tools.
+ *
+ * for new version cos file system is the shell which using the head bucket
+ * to decide whether normal bucket or posix bucket, if it is posix bucket
+ * direct use the network shell implements.
  */
 @InterfaceAudience.Private
 @InterfaceStability.Stable
@@ -51,31 +53,25 @@ public class CosFileSystem extends FileSystem {
 
     static final String SCHEME = "cosn";
     static final String PATH_DELIMITER = Path.SEPARATOR;
-
     static final Charset METADATA_ENCODING = StandardCharsets.UTF_8;
-    // The length of name:value pair should be less than or equal to 1024 bytes.
-    static final int MAX_XATTR_SIZE = 1024;
 
     private URI uri;
     private String bucket;
-    private boolean isMergeBucket;
-    private boolean checkMergeBucket;
-    private int mergeBucketMaxListNum;
-    private int normalBucketMaxListNum;
-    private NativeFileSystemStore store;
+    private String formatBucket;
     private Path workingDir;
+    private boolean isPosixBucket;
+    private boolean isGatewayMode;
+    private boolean isCosProcess;
+    private NativeFileSystemStore store;
+
     private String owner = "Unknown";
     private String group = "Unknown";
-
-    private ExecutorService boundedIOThreadPool;
-    private ExecutorService boundedCopyThreadPool;
-
     private UserGroupInformation currentUser;
-
     private boolean enableRangerPluginPermissionCheck = false;
     public static RangerQcloudObjectStorageClient rangerQcloudObjectStorageStorageClient = null;
 
-    private String formatBucket;
+    private FileSystem actualImplFS = null;
+
 
     public CosFileSystem() {
     }
@@ -102,6 +98,7 @@ public class CosFileSystem extends FileSystem {
         UserGroupInformation.setConfiguration(conf);
         this.currentUser = UserGroupInformation.getCurrentUser();
 
+        // use the topper cos ranger. ignore the ofs ranger
         initRangerClientImpl(conf);
 
         this.bucket = uri.getHost();
@@ -111,6 +108,7 @@ public class CosFileSystem extends FileSystem {
             this.store = createDefaultStore(conf);
         }
 
+        // this store no need to init twice.
         this.store.initialize(uri, conf);
         this.uri = URI.create(uri.getScheme() + "://" + uri.getAuthority());
         this.workingDir =
@@ -121,118 +119,59 @@ public class CosFileSystem extends FileSystem {
         LOG.debug("uri: {}, bucket: {}, working dir: {}, owner: {}, group: {}.\n" +
                         "configuration: {}.",
                 uri, bucket, workingDir, owner, group, conf);
-        BufferPool.getInstance().initialize(getConf());
 
-        // head the bucket to judge whether the merge bucket
-        this.isMergeBucket = false;
-        this.checkMergeBucket = this.getConf().getBoolean(
-                CosNConfigKeys.OPEN_CHECK_MERGE_BUCKET,
-                CosNConfigKeys.DEFAULT_CHECK_MERGE_BUCKET
-        );
+        this.isPosixBucket = false;
+        this.isCosProcess = false;
+        this.isGatewayMode = this.getConf().getBoolean(
+                CosNConfigKeys.COSN_POSIX_BUCKET_GATEWAY_MODE_ENABLED,
+                CosNConfigKeys.DEFAULT_COSN_POSIX_BUCKET_GATEWAY_MODE_ENABLED);
 
-        if (this.checkMergeBucket) { // control
-            HeadBucketResult headBucketResult = this.store.headBucket(this.bucket);
-            int mergeBucketMaxListNum = this.getConf().getInt(
-                    CosNConfigKeys.MERGE_BUCKET_MAX_LIST_NUM,
-                    CosNConfigKeys.DEFAULT_MERGE_BUCKET_MAX_LIST_NUM
-            );
-            if (headBucketResult.isMergeBucket()) {
-                this.isMergeBucket = true;
-                this.mergeBucketMaxListNum = mergeBucketMaxListNum;
-                this.store.setMergeBucket(true);
-            }
+        String PosixBucketNoGatewayImpl = this.getConf().get(
+                CosNConfigKeys.COSN_POSIX_BUCKET_NOGATEWAY_IMPL,
+                CosNConfigKeys.DEFAULT_COSN_POSIX_BUCKET_NOGATEWAY_IMPL);
+
+        // head the bucket to judge whether the posix bucket
+        HeadBucketResult headBucketResult = this.store.headBucket(this.bucket);
+        this.isPosixBucket = headBucketResult.isMergeBucket();
+
+        LOG.info("cos file system bucket is posix bucket {}, is gateway mode {}",
+                isPosixBucket, isGatewayMode);
+        if (isPosixBucket && !isGatewayMode) { // ofs implements
+            // network version start from the 2.7
+            // sdk version start from the 1.0.4
+            this.actualImplFS = getActualFileSystemByClassName(PosixBucketNoGatewayImpl);
+            this.isCosProcess = false;
+            this.store.setPosixBucket(true);
+            // before the init, must transfer the config and disable the range in ofs
+            this.transferOfsConfig();
+        } else { // normal cos hadoop file system implements
+            this.isCosProcess = true;
+            this.actualImplFS = getActualFileSystemByClassName("org.apache.hadoop.fs.CosHadoopFileSystem");
+            ((CosHadoopFileSystem) this.actualImplFS).setNativeFileSystemStore(this.store, this.bucket, this.uri,
+                    this.owner, this.group, isPosixBucket);
+            this.actualImplFS.setWorkingDirectory(this.workingDir);
         }
-        LOG.info("cos file system bucket is merged {}", this.isMergeBucket);
-        this.normalBucketMaxListNum = this.getConf().getInt(
-                CosNConfigKeys.NORMAL_BUCKET_MAX_LIST_NUM,
-                CosNConfigKeys.DEFAULT_NORMAL_BUCKET_MAX_LIST_NUM);
 
-        // initialize the thread pool
-        int uploadThreadPoolSize = this.getConf().getInt(
-                CosNConfigKeys.UPLOAD_THREAD_POOL_SIZE_KEY,
-                CosNConfigKeys.DEFAULT_UPLOAD_THREAD_POOL_SIZE
-        );
-        int readAheadPoolSize = this.getConf().getInt(
-                CosNConfigKeys.READ_AHEAD_QUEUE_SIZE,
-                CosNConfigKeys.DEFAULT_READ_AHEAD_QUEUE_SIZE
-        );
-        Preconditions.checkArgument(uploadThreadPoolSize > 0,
-                String.format("The uploadThreadPoolSize[%d] should be positive.", uploadThreadPoolSize));
-        Preconditions.checkArgument(readAheadPoolSize > 0,
-                String.format("The readAheadQueueSize[%d] should be positive.", readAheadPoolSize));
-        // 核心线程数取用户配置的为准，最大线程数结合用户配置和IO密集型任务的最优线程数来看
-        int ioCoreTaskSize = uploadThreadPoolSize + readAheadPoolSize;
-        int ioMaxTaskSize = Math.max(uploadThreadPoolSize + readAheadPoolSize,
-                Runtime.getRuntime().availableProcessors() * 2 + 1);
-        if (this.getConf().get(CosNConfigKeys.IO_THREAD_POOL_MAX_SIZE_KEY) != null) {
-            int ioThreadPoolMaxSize = this.getConf().getInt(
-                    CosNConfigKeys.IO_THREAD_POOL_MAX_SIZE_KEY, CosNConfigKeys.DEFAULT_IO_THREAD_POOL_MAX_SIZE);
-            Preconditions.checkArgument(ioThreadPoolMaxSize > 0,
-                    String.format("The ioThreadPoolMaxSize[%d] should be positive.", ioThreadPoolMaxSize));
-            // 如果设置了 IO 线程池的最大限制，则整个线程池需要被限制住
-            ioCoreTaskSize = Math.min(ioCoreTaskSize, ioThreadPoolMaxSize);
-            ioMaxTaskSize = ioThreadPoolMaxSize;
+        if (this.actualImplFS == null) {
+            // should never reach here
+            throw new IOException("impl file system is null");
         }
-        long threadKeepAlive = this.getConf().getLong(
-                CosNConfigKeys.THREAD_KEEP_ALIVE_TIME_KEY,
-                CosNConfigKeys.DEFAULT_THREAD_KEEP_ALIVE_TIME);
-        Preconditions.checkArgument(threadKeepAlive > 0,
-                String.format("The threadKeepAlive [%d] should be positive.", threadKeepAlive));
-        this.boundedIOThreadPool = new ThreadPoolExecutor(
-                ioCoreTaskSize, ioMaxTaskSize,
-                threadKeepAlive, TimeUnit.SECONDS,
-                new LinkedBlockingQueue<Runnable>(ioCoreTaskSize),
-                new ThreadFactoryBuilder().setNameFormat("cos-transfer-shared-%d").setDaemon(true).build(),
-                new RejectedExecutionHandler() {
-                    @Override
-                    public void rejectedExecution(Runnable r,
-                                                  ThreadPoolExecutor executor) {
-                        if (!executor.isShutdown()) {
-                            try {
-                                executor.getQueue().put(r);
-                            } catch (InterruptedException e) {
-                                LOG.error("put a io task into the download " +
-                                        "thread pool occurs an exception.", e);
-                                throw new RejectedExecutionException(
-                                        "Putting the io task failed due to the interruption", e);
-                            }
-                        } else {
-                            LOG.error("The bounded io thread pool has been shutdown.");
-                            throw new RejectedExecutionException("The bounded io thread pool has been shutdown");
-                        }
-                    }
-                }
-        );
-
-        int copyThreadPoolSize = this.getConf().getInt(
-                CosNConfigKeys.COPY_THREAD_POOL_SIZE_KEY,
-                CosNConfigKeys.DEFAULT_COPY_THREAD_POOL_SIZE
-        );
-        Preconditions.checkArgument(copyThreadPoolSize > 0,
-                String.format("The copyThreadPoolSize[%d] should be positive.", copyThreadPoolSize));
-        this.boundedCopyThreadPool = new ThreadPoolExecutor(
-                copyThreadPoolSize, copyThreadPoolSize,
-                threadKeepAlive, TimeUnit.SECONDS,
-                new LinkedBlockingQueue<Runnable>(copyThreadPoolSize),
-                new ThreadFactoryBuilder().setNameFormat("cos-copy-%d").setDaemon(true).build(),
-                new RejectedExecutionHandler() {
-                    @Override
-                    public void rejectedExecution(Runnable r,
-                                                  ThreadPoolExecutor executor) {
-                        if (!executor.isShutdown()) {
-                            try {
-                                executor.getQueue().put(r);
-                            } catch (InterruptedException e) {
-                                LOG.error("put a copy task into the download " +
-                                        "thread pool occurs an exception.", e);
-                            }
-                        }
-                    }
-                }
-        );
+        this.actualImplFS.initialize(uri, conf);
     }
 
-    private static NativeFileSystemStore createDefaultStore(Configuration conf) {
+    // load class to get relate file system
+    private static FileSystem getActualFileSystemByClassName(String className) throws IOException {
+        try {
+            Class<?> actualClass = Class.forName(className);
+            return (FileSystem)actualClass.newInstance();
+        } catch (ClassNotFoundException | InstantiationException | IllegalAccessException e) {
+            String errMsg = String.format("load class failed, className: %s", className);
+            LOG.error(errMsg, e);
+            throw new IOException(errMsg, e);
+        }
+    }
+
+    public static NativeFileSystemStore createDefaultStore(Configuration conf) {
         NativeFileSystemStore store = new CosNativeFileSystemStore();
         RetryPolicy basePolicy =
                 RetryPolicies.retryUpToMaximumCountWithFixedSleep(
@@ -322,36 +261,6 @@ public class CosFileSystem extends FileSystem {
         return ownerInfoId;
     }
 
-    private static String pathToKey(Path path) {
-        if (path.toUri().getScheme() != null && path.toUri().getPath().isEmpty()) {
-            // allow uris without trailing slash after bucket to refer to root,
-            // like cosn://mybucket
-            return "";
-        }
-        if (!path.isAbsolute()) {
-            throw new IllegalArgumentException("Path must be absolute: " + path);
-        }
-        String ret = path.toUri().getPath();
-        if (ret.endsWith("/") && (ret.indexOf("/") != ret.length() - 1)) {
-            ret = ret.substring(0, ret.length() - 1);
-        }
-        return ret;
-    }
-
-    private static Path keyToPath(String key) {
-        if (!key.startsWith(PATH_DELIMITER)) {
-            return new Path("/" + key);
-        } else {
-            return new Path(key);
-        }
-    }
-
-    private Path makeAbsolute(Path path) {
-        if (path.isAbsolute()) {
-            return path;
-        }
-        return new Path(workingDir, path);
-    }
 
     @Override
     public Path getHomeDirectory() {
@@ -364,12 +273,22 @@ public class CosFileSystem extends FileSystem {
     }
 
     /**
-     * This optional operation is not yet supported.
+     * This optional operation is not yet supported for current cos process.
+     * but it can be used by posix bucket.
      */
     @Override
     public FSDataOutputStream append(Path f, int bufferSize,
-                                     Progressable progress) {
-        throw new UnsupportedOperationException("Not supported currently");
+                                     Progressable progress) throws IOException {
+        LOG.debug("append file [{}] in COS.", f);
+        checkPermission(f, RangerAccessType.WRITE);
+        return this.actualImplFS.append(f, bufferSize, progress);
+    }
+
+    @Override
+    public boolean truncate(Path f, long newLength) throws IOException {
+        LOG.debug("truncate file [{}] in COS.", f);
+        checkPermission(f, RangerAccessType.WRITE);
+        return this.actualImplFS.truncate(f, newLength);
     }
 
     @Override
@@ -378,204 +297,24 @@ public class CosFileSystem extends FileSystem {
                                      int bufferSize, short replication,
                                      long blockSize, Progressable progress)
             throws IOException {
-        // check permission
-        checkPermission(f, RangerAccessType.WRITE);
-
-        // preconditions
-        try {
-            FileStatus targetFileStatus = this.getFileStatus(f);
-            if (null != targetFileStatus && !overwrite) {
-                throw new FileAlreadyExistsException("File already exists: " + f);
-            }
-            if (null != targetFileStatus && targetFileStatus.isDirectory()) {
-                throw new FileAlreadyExistsException("Directory already exists: " + f);
-            }
-        } catch (FileNotFoundException e) {
-            validatePath(f);
-        }
-
         LOG.debug("Creating a new file [{}] in COS.", f);
-
-        Path absolutePath = makeAbsolute(f);
-        String key = pathToKey(absolutePath);
-        long uploadPartSize = this.getConf().getLong(
-                CosNConfigKeys.COSN_UPLOAD_PART_SIZE_KEY, CosNConfigKeys.DEFAULT_UPLOAD_PART_SIZE);
-        boolean uploadChecksEnabled = this.getConf().getBoolean(CosNConfigKeys.COSN_UPLOAD_CHECKS_ENABLE_KEY,
-                CosNConfigKeys.DEFAULT_COSN_UPLOAD_CHECKS_ENABLE);
-        return new FSDataOutputStream(
-                new CosNFSDataOutputStream(getConf(), store, key,
-                        uploadPartSize,
-                        this.boundedIOThreadPool, uploadChecksEnabled), statistics);
+        checkPermission(f, RangerAccessType.WRITE);
+        return this.actualImplFS.create(f, permission, overwrite, bufferSize,
+                replication, blockSize, progress);
     }
 
-    private boolean rejectRootDirectoryDelete(boolean isEmptyDir,
-                                              boolean recursive)
-            throws PathIOException {
-        if (isEmptyDir) {
-            return true;
-        }
-        if (recursive) {
-            return false;
-        } else {
-            throw new PathIOException(this.bucket, "Can not delete root path");
-        }
-    }
 
     @Override
     public boolean delete(Path f, boolean recursive) throws IOException {
         LOG.debug("Ready to delete path: {}. recursive: {}.", f, recursive);
-
         checkPermission(f, RangerAccessType.DELETE);
-        FileStatus status;
-        try {
-            status = getFileStatus(f);
-        } catch (FileNotFoundException e) {
-            LOG.debug("Delete called for '{}', but the file does not exist and returning the false.", f);
-            return false;
-        }
-
-        Path absolutePath = makeAbsolute(f);
-        String key = pathToKey(absolutePath);
-        if (key.compareToIgnoreCase("/") == 0) {
-            FileStatus[] fileStatuses = listStatus(f);
-            return this.rejectRootDirectoryDelete(fileStatuses.length == 0,
-                    recursive);
-        }
-
-        if (status.isDirectory()) {
-            if (!key.endsWith(PATH_DELIMITER)) {
-                key += PATH_DELIMITER;
-            }
-            if (!recursive && listStatus(f).length > 0) {
-                throw new IOException("Can not delete " + f
-                        + " as is a not empty directory and recurse option is" +
-                        " false");
-            }
-            // how to tell the result
-            if (!isMergeBucket) {
-                internalRecursiveDelete(key);
-            } else {
-                internalAutoRecursiveDelete(key);
-            }
-        } else {
-            LOG.debug("Delete the cos key [{}].", key);
-            store.delete(key);
-        }
-
-        if (!isMergeBucket) {
-            createParentDirectoryIfNecessary(f);
-        }
-        return true;
-    }
-
-
-    private void internalRecursiveDelete(String key) throws IOException {
-        CosNDeleteFileContext deleteFileContext = new CosNDeleteFileContext();
-        int deleteToFinishes = 0;
-
-        String priorLastKey = null;
-        do {
-            CosNPartialListing listing =
-                    store.list(key, this.normalBucketMaxListNum, priorLastKey, true);
-            for (FileMetadata file : listing.getFiles()) {
-                this.boundedCopyThreadPool.execute(new CosNDeleteFileTask(
-                        this.store, file.getKey(), deleteFileContext));
-                deleteToFinishes++;
-                if (!deleteFileContext.isDeleteSuccess()) {
-                    break;
-                }
-            }
-            for (FileMetadata commonPrefix : listing.getCommonPrefixes()) {
-                this.boundedCopyThreadPool.execute(new CosNDeleteFileTask(
-                        this.store, commonPrefix.getKey(), deleteFileContext));
-                deleteToFinishes++;
-                if (!deleteFileContext.isDeleteSuccess()) {
-                    break;
-                }
-            }
-            priorLastKey = listing.getPriorLastKey();
-        } while (priorLastKey != null);
-
-        deleteFileContext.lock();
-        try {
-            deleteFileContext.awaitAllFinish(deleteToFinishes);
-        } catch (InterruptedException e) {
-            LOG.warn("interrupted when wait delete to finish");
-        } finally {
-            deleteFileContext.unlock();
-        }
-
-        // according the flag and exception in thread opr to throw this out
-        if (!deleteFileContext.isDeleteSuccess() && deleteFileContext.hasException()) {
-            throw deleteFileContext.getIOException();
-        }
-
-        try {
-            LOG.debug("Delete the cos key [{}].", key);
-            store.delete(key);
-        } catch (Exception e) {
-            LOG.error("Delete the key failed.");
-        }
-    }
-
-    // use by merge bucket which support recursive delete dirs by setting flag parameter
-    private void internalAutoRecursiveDelete(String key) throws IOException {
-        LOG.debug("Delete the cos key auto recursive [{}].", key);
-        store.deleteRecursive(key);
+        return this.actualImplFS.delete(f, recursive);
     }
 
     @Override
     public FileStatus getFileStatus(Path f) throws IOException {
         LOG.debug("Get file status: {}.", f);
-        Path absolutePath = makeAbsolute(f);
-        String key = pathToKey(absolutePath);
-
-        if (key.length() == 0 || key.equals(PATH_DELIMITER)) { // root always exists
-            return newDirectory(absolutePath);
-        }
-
-        CosNResultInfo headInfo = new CosNResultInfo();
-        FileMetadata meta = store.retrieveMetadata(key, headInfo);
-        if (meta != null) {
-            if (meta.isFile()) {
-                LOG.debug("Retrieve the cos key [{}] to find that it is a file.", key);
-                return newFile(meta, absolutePath);
-            } else {
-                LOG.debug("Retrieve the cos key [{}] to find that it is a directory.", key);
-                return newDirectory(meta, absolutePath);
-            }
-        }
-
-        if (isMergeBucket) {
-            throw new FileNotFoundException("No such file or directory in merge bucket'" + absolutePath + "'");
-        }
-
-        if (!key.endsWith(PATH_DELIMITER)) {
-            key += PATH_DELIMITER;
-        }
-
-        int maxKeys = this.getConf().getInt(
-                CosNConfigKeys.FILESTATUS_LIST_MAX_KEYS,
-                CosNConfigKeys.DEFAULT_FILESTATUS_LIST_MAX_KEYS
-        );
-
-        LOG.debug("List the cos key [{}] to judge whether it is a directory or not. max keys [{}]", key, maxKeys);
-        CosNResultInfo listInfo = new CosNResultInfo();
-        CosNPartialListing listing = store.list(key, maxKeys, listInfo);
-        if (listing.getFiles().length > 0 || listing.getCommonPrefixes().length > 0) {
-            LOG.debug("List the cos key [{}] to find that it is a directory.", key);
-            return newDirectory(absolutePath);
-        }
-
-        if (listInfo.isKeySameToPrefix()) {
-            LOG.info("List the cos key [{}] same to prefix, head-id:[{}], " +
-                            "list-id:[{}], list-type:[{}], thread-id:[{}], thread-name:[{}]",
-                    key, headInfo.getRequestID(), listInfo.getRequestID(),
-                    listInfo.isKeySameToPrefix(), Thread.currentThread().getId(), Thread.currentThread().getName());
-        }
-        LOG.debug("Can not find the cos key [{}] on COS.", key);
-
-        throw new FileNotFoundException("No such file or directory '" + absolutePath + "'");
+        return this.actualImplFS.getFileStatus(f);
     }
 
     @Override
@@ -597,109 +336,7 @@ public class CosFileSystem extends FileSystem {
     public FileStatus[] listStatus(Path f) throws FileNotFoundException, IOException {
         LOG.debug("list status:" + f);
         checkPermission(f, RangerAccessType.LIST);
-
-        Path absolutePath = makeAbsolute(f);
-        String key = pathToKey(absolutePath);
-        int listMaxLength = this.normalBucketMaxListNum;
-        if (checkMergeBucket && isMergeBucket) {
-            listMaxLength = this.mergeBucketMaxListNum;
-        }
-
-
-        if (key.length() > 0) {
-            FileMetadata meta = store.retrieveMetadata(key);
-            if (meta != null && meta.isFile()) {
-                return new FileStatus[]{newFile(meta, absolutePath)};
-            }
-        }
-
-        if (!key.endsWith(PATH_DELIMITER)) {
-            key += PATH_DELIMITER;
-        }
-
-        // if introduce the get file status to solve the file not exist exception
-        // will expand the list qps, for now only used get file status for merge bucket.
-        if (this.isMergeBucket) {
-            try {
-                this.getFileStatus(f);
-            } catch (FileNotFoundException e) {
-                throw new FileNotFoundException("No such file or directory:" + f);
-            }
-        }
-
-        URI pathUri = absolutePath.toUri();
-        Set<FileStatus> status = new TreeSet<FileStatus>();
-        String priorLastKey = null;
-        do {
-            CosNPartialListing listing = store.list(key, listMaxLength,
-                    priorLastKey, false);
-            for (FileMetadata fileMetadata : listing.getFiles()) {
-                Path subPath = keyToPath(fileMetadata.getKey());
-                if (fileMetadata.getKey().equals(key)) {
-                    // this is just the directory we have been asked to list
-                    LOG.debug("This is just the directory we have been asked to list. cos key: {}.",
-                            fileMetadata.getKey());
-                } else {
-                    status.add(newFile(fileMetadata, subPath));
-                }
-            }
-            for (FileMetadata commonPrefix : listing.getCommonPrefixes()) {
-                Path subpath = keyToPath(commonPrefix.getKey());
-                String relativePath =
-                        pathUri.relativize(subpath.toUri()).getPath();
-                status.add(newDirectory(commonPrefix, new Path(absolutePath,
-                        relativePath)));
-            }
-            priorLastKey = listing.getPriorLastKey();
-        } while (priorLastKey != null);
-
-        return status.toArray(new FileStatus[status.size()]);
-    }
-
-    private FileStatus newFile(FileMetadata meta, Path path) {
-        return new CosNFileStatus(meta.getLength(), false, 1, getDefaultBlockSize(),
-                meta.getLastModified(), 0, null, this.owner, this.group,
-                path.makeQualified(this.getUri(), this.getWorkingDirectory()), meta.getETag(), meta.getCrc64ecm(),
-                meta.getVersionId());
-    }
-
-    private FileStatus newDirectory(Path path) {
-        return new CosNFileStatus(0, true, 1, 0, 0, 0, null, this.owner, this.group,
-                path.makeQualified(this.getUri(), this.getWorkingDirectory()));
-    }
-
-    private FileStatus newDirectory(FileMetadata meta, Path path) {
-        if (meta == null) {
-            return newDirectory(path);
-        }
-        return new CosNFileStatus(0, true, 1, 0,
-                meta.getLastModified(), 0, null, this.owner, this.group,
-                path.makeQualified(this.getUri(), this.getWorkingDirectory()), meta.getETag(), meta.getCrc64ecm(),
-                meta.getVersionId());
-    }
-
-    /**
-     * Validate the path from the bottom up.
-     *
-     * @param path the absolute path to check.
-     * @throws IOException an IOException occurred when getting the path's metadata.
-     */
-    private void validatePath(Path path) throws IOException {
-        Path parent = path.getParent();
-        do {
-            try {
-                FileStatus parentFileStatus = getFileStatus(parent);
-                if (null != parentFileStatus && parentFileStatus.isDirectory()) {
-                    break;
-                }
-                if (null != parentFileStatus && !parentFileStatus.isDirectory()) {
-                    throw new ParentNotDirectoryException(String.format(
-                            "Can't create under the path '%s', because it is a file.", parent));
-                }
-            } catch (FileNotFoundException ignored) {
-            }
-            parent = parent.getParent();
-        } while (parent != null);
+        return this.actualImplFS.listStatus(f);
     }
 
     @Override
@@ -707,317 +344,27 @@ public class CosFileSystem extends FileSystem {
             throws IOException {
         LOG.debug("mkdirs path: {}.", f);
         checkPermission(f, RangerAccessType.WRITE);
-        try {
-            FileStatus fileStatus = getFileStatus(f);
-            if (fileStatus.isDirectory()) {
-                return true;
-            } else {
-                throw new FileAlreadyExistsException("Path is a file: " + f);
-            }
-        } catch (FileNotFoundException e) {
-            validatePath(f);
-            boolean result;
-            if (isMergeBucket) {
-                result = mkDirAutoRecursively(f, permission);
-            } else {
-                result = mkDirRecursively(f, permission);
-            }
-            return result;
-        }
-    }
-
-    /**
-     * Create a directory recursively.
-     *
-     * @param f          Absolute path to the directory
-     * @param permission Directory permissions
-     * @return Return true if the creation was successful,  throw a IOException.
-     * @throws IOException An IOException occurred when creating a directory object on COS.
-     */
-    public boolean mkDirRecursively(Path f, FsPermission permission)
-            throws IOException {
-        LOG.debug("Make the directory recursively. Path: {}, FsPermission: {}.", f, permission);
-        Path absolutePath = makeAbsolute(f);
-        List<Path> paths = new ArrayList<Path>();
-        do {
-            paths.add(absolutePath);
-            absolutePath = absolutePath.getParent();
-        } while (absolutePath != null);
-
-        for (Path path : paths) {
-            if (path.isRoot()) {
-                break;
-            }
-            try {
-                FileStatus fileStatus = getFileStatus(path);
-                if (fileStatus.isFile()) {
-                    throw new FileAlreadyExistsException(
-                            String.format(
-                                    "Can't make directory for path '%s' since" +
-                                            " it is a file.", f));
-                }
-                if (fileStatus.isDirectory()) {
-                    if (fileStatus.getModificationTime() > 0) {
-                        break;
-                    } else {
-                        throw new FileNotFoundException("Dir '" + path + "' doesn't exist in COS");
-                    }
-                }
-            } catch (FileNotFoundException e) {
-                LOG.debug("Make the directory [{}] on COS.", path);
-                String folderPath = pathToKey(makeAbsolute(path));
-                if (!folderPath.endsWith(PATH_DELIMITER)) {
-                    folderPath += PATH_DELIMITER;
-                }
-                store.storeEmptyFile(folderPath);
-            }
-        }
-        return true;
-    }
-
-    /**
-     * Create a directory recursively auto by merge plan
-     * which the put object interface decide the dir which end with the "/"
-     *
-     * @param f          Absolute path to the directory
-     * @param permission Directory permissions
-     * @return Return true if the creation was successful,  throw a IOException.
-     * @throws IOException An IOException occurred when creating a directory object on COS.
-     */
-    public boolean mkDirAutoRecursively(Path f, FsPermission permission)
-            throws IOException {
-        LOG.debug("Make the directory recursively auto. Path: {}, FsPermission: {}.", f, permission);
-        // add the end '/' to the path which server auto create middle dir
-        String folderPath = pathToKey(makeAbsolute(f));
-        if (!folderPath.endsWith(PATH_DELIMITER)) {
-            folderPath += PATH_DELIMITER;
-        }
-        store.storeEmptyFile(folderPath);
-        return true;
+        return this.actualImplFS.mkdirs(f, permission);
     }
 
     @Override
     public FSDataInputStream open(Path f, int bufferSize) throws IOException {
-        checkPermission(f, RangerAccessType.READ);
         LOG.debug("Open file [{}] to read, buffer [{}]", f, bufferSize);
-
-        FileStatus fileStatus = getFileStatus(f); // will throw if the file doesn't
-        // exist
-        if (fileStatus.isDirectory()) {
-            throw new FileNotFoundException("'" + f + "' is a directory");
-        }
-        LOG.info("Opening '" + f + "' for reading");
-        Path absolutePath = makeAbsolute(f);
-        String key = pathToKey(absolutePath);
-        return new FSDataInputStream(new BufferedFSInputStream(
-                new CosNFSInputStream(this.getConf(), store, statistics, key,
-                        fileStatus.getLen(), this.boundedIOThreadPool),
-                bufferSize));
+        checkPermission(f, RangerAccessType.READ);
+        return this.actualImplFS.open(f, bufferSize);
     }
 
     @Override
     public boolean rename(Path src, Path dst) throws IOException {
         LOG.debug("Rename the source path [{}] to the dest path [{}].", src, dst);
-
         checkPermission(src, RangerAccessType.DELETE);
         checkPermission(dst, RangerAccessType.WRITE);
-
-        // Renaming the root directory is not allowed
-        if (src.isRoot()) {
-            LOG.debug("Cannot rename the root directory of a filesystem.");
-            return false;
-        }
-
-        // check the source path whether exists or not, if not return false.
-        FileStatus srcFileStatus;
-        try {
-            srcFileStatus = this.getFileStatus(src);
-        } catch (FileNotFoundException e) {
-            LOG.debug("The source path [{}] is not exist.", src);
-            return false;
-        }
-
-        // Source path and destination path are not allowed to be the same
-        if (src.equals(dst)) {
-            LOG.debug("The source path and the dest path refer to the same file or " +
-                    "directory: {}", dst);
-            throw new IOException("the source path and dest path refer to the " +
-                    "same file or directory");
-        }
-
-        // It is not allowed to rename a parent directory to its subdirectory
-        Path dstParentPath;
-        dstParentPath = dst.getParent();
-        while (null != dstParentPath && !src.equals(dstParentPath)) {
-            dstParentPath = dstParentPath.getParent();
-        }
-        if (null != dstParentPath) {
-            LOG.debug("It is not allowed to rename a parent directory:{} to " +
-                    "its subdirectory:{}.", src, dst);
-            throw new IOException(String.format(
-                    "It is not allowed to rename a parent directory:%s to its" +
-                            " subdirectory:%s",
-                    src, dst));
-        }
-
-        FileStatus dstFileStatus = null;
-        try {
-            dstFileStatus = this.getFileStatus(dst);
-
-            // The destination path exists and is a file,
-            // and the rename operation is not allowed.
-            //
-            if (dstFileStatus.isFile()) {
-//                throw new FileAlreadyExistsException(String.format(
-//                        "File:%s already exists", dstFileStatus.getPath()));
-                LOG.debug("File: {} already exists.", dstFileStatus.getPath());
-                return false;
-            } else {
-                // The destination path is an existing directory,
-                // and it is checked whether there is a file or directory
-                // with the same name as the source path under the
-                // destination path
-                dst = new Path(dst, src.getName());
-                FileStatus[] statuses;
-                try {
-                    statuses = this.listStatus(dst);
-                } catch (FileNotFoundException e) {
-                    statuses = null;
-                }
-                if (null != statuses && statuses.length > 0) {
-                    LOG.debug("Cannot rename {} to {}, file already exists.",
-                            src, dst);
-//                    throw new FileAlreadyExistsException(
-//                            String.format(
-//                                    "File: %s already exists", dst
-//                            )
-//                    );
-                    return false;
-                }
-            }
-        } catch (FileNotFoundException e) {
-            // destination path not exists
-            Path tempDstParentPath = dst.getParent();
-            FileStatus dstParentStatus = this.getFileStatus(tempDstParentPath);
-            if (!dstParentStatus.isDirectory()) {
-                throw new IOException(String.format(
-                        "Cannot rename %s to %s, %s is a file", src, dst, dst.getParent()
-                ));
-            }
-            // The default root directory is definitely there.
-        }
-
-        if (!isMergeBucket) {
-            return internalCopyAndDelete(src, dst, srcFileStatus.isDirectory());
-        } else {
-            return internalRename(src, dst);
-        }
-    }
-
-    private boolean internalCopyAndDelete(Path srcPath, Path dstPath, boolean isDir) throws IOException {
-        boolean result = false;
-        if (isDir) {
-            result = this.copyDirectory(srcPath, dstPath);
-        } else {
-            result = this.copyFile(srcPath, dstPath);
-        }
-
-        if (!result) {
-            //Since rename is a non-atomic operation, after copy fails,
-            // it is not allowed to delete the data of the original path to
-            // ensure data security.
-            return false;
-        } else {
-            return this.delete(srcPath, true);
-        }
-    }
-
-    private boolean internalRename(Path srcPath, Path dstPath) throws IOException {
-        String srcKey = pathToKey(srcPath);
-        String dstKey = pathToKey(dstPath);
-        this.store.rename(srcKey, dstKey);
-        return true;
-    }
-
-    private boolean copyFile(Path srcPath, Path dstPath) throws IOException {
-        String srcKey = pathToKey(srcPath);
-        String dstKey = pathToKey(dstPath);
-        this.store.copy(srcKey, dstKey);
-        return true;
-    }
-
-    private boolean copyDirectory(Path srcPath, Path dstPath) throws IOException {
-        String srcKey = pathToKey(srcPath);
-        if (!srcKey.endsWith(PATH_DELIMITER)) {
-            srcKey += PATH_DELIMITER;
-        }
-        String dstKey = pathToKey(dstPath);
-        if (!dstKey.endsWith(PATH_DELIMITER)) {
-            dstKey += PATH_DELIMITER;
-        }
-
-        if (dstKey.startsWith(srcKey)) {
-            throw new IOException("can not copy a directory to a subdirectory" +
-                    " of self");
-        }
-
-        if (this.store.retrieveMetadata(srcKey) == null) {
-            this.store.storeEmptyFile(srcKey);
-        } else {
-            this.store.copy(srcKey, dstKey);
-        }
-
-        CosNCopyFileContext copyFileContext = new CosNCopyFileContext();
-
-        int copiesToFinishes = 0;
-        String priorLastKey = null;
-        do {
-            CosNPartialListing objectList = this.store.list(srcKey,
-                    this.normalBucketMaxListNum, priorLastKey, true);
-            for (FileMetadata file : objectList.getFiles()) {
-                this.boundedCopyThreadPool.execute(new CosNCopyFileTask(
-                        this.store,
-                        file.getKey(),
-                        dstKey.concat(file.getKey().substring(srcKey.length())),
-                        copyFileContext));
-                copiesToFinishes++;
-                if (!copyFileContext.isCopySuccess()) {
-                    break;
-                }
-            }
-            priorLastKey = objectList.getPriorLastKey();
-        } while (null != priorLastKey);
-
-        copyFileContext.lock();
-        try {
-            copyFileContext.awaitAllFinish(copiesToFinishes);
-        } catch (InterruptedException e) {
-            LOG.warn("interrupted when wait copies to finish");
-        } finally {
-            copyFileContext.unlock();
-        }
-        return copyFileContext.isCopySuccess();
-    }
-
-    private void createParentDirectoryIfNecessary(Path path) throws IOException {
-        Path parent = path.getParent();
-        if (null != parent && !parent.isRoot()) {
-            String parentKey = pathToKey(parent);
-            if (!StringUtils.isNullOrEmpty(parentKey) && !exists(parent)) {
-                LOG.debug("Create a parent directory [{}] for the path [{}].", parent, path);
-                if (!parentKey.endsWith("/")) {
-                    parentKey += "/";
-                }
-                store.storeEmptyFile(parentKey);
-            }
-        }
-    }
+        return this.actualImplFS.rename(src, dst);
+	}
 
     @Override
     public long getDefaultBlockSize() {
-        return getConf().getLong(
-                CosNConfigKeys.COSN_BLOCK_SIZE_KEY,
-                CosNConfigKeys.DEFAULT_BLOCK_SIZE);
+        return this.actualImplFS.getDefaultBlockSize();
     }
 
     /**
@@ -1093,37 +440,12 @@ public class CosFileSystem extends FileSystem {
 
     @Override
     public FileChecksum getFileChecksum(Path f, long length) throws IOException {
-        Preconditions.checkArgument(length >= 0);
         LOG.debug("call the checksum for the path: {}.", f);
+        Preconditions.checkArgument(length >= 0);
 
         // The order of each file, must support both crc at same time, how to tell the difference crc request?
         checkPermission(f, RangerAccessType.READ);
-
-        if (this.getConf().getBoolean(CosNConfigKeys.CRC64_CHECKSUM_ENABLED,
-                CosNConfigKeys.DEFAULT_CRC64_CHECKSUM_ENABLED)) {
-            Path absolutePath = makeAbsolute(f);
-            String key = pathToKey(absolutePath);
-
-            FileMetadata fileMetadata = this.store.retrieveMetadata(key);
-            if (null == fileMetadata) {
-                throw new FileNotFoundException("File or directory doesn't exist: " + f);
-            }
-            String crc64ecm = fileMetadata.getCrc64ecm();
-            return crc64ecm != null ? new CRC64Checksum(crc64ecm) : super.getFileChecksum(f, length);
-        } else if (this.getConf().getBoolean(CosNConfigKeys.CRC32C_CHECKSUM_ENABLED,
-                CosNConfigKeys.DEFAULT_CRC32C_CHECKSUM_ENABLED)) {
-            Path absolutePath = makeAbsolute(f);
-            String key = pathToKey(absolutePath);
-            FileMetadata fileMetadata = this.store.retrieveMetadata(key);
-            if (null == fileMetadata) {
-                throw new FileNotFoundException("File or directory doesn't exist: " + f);
-            }
-            String crc32cm = fileMetadata.getCrc32cm();
-            return crc32cm != null ? new CRC32CCheckSum(crc32cm) : super.getFileChecksum(f, length);
-        } else {
-            // disabled
-            return super.getFileChecksum(f, length);
-        }
+        return this.actualImplFS.getFileChecksum(f, length);
     }
 
 
@@ -1139,31 +461,8 @@ public class CosFileSystem extends FileSystem {
     @Override
     public void setXAttr(Path f, String name, byte[] value, EnumSet<XAttrSetFlag> flag) throws IOException {
         LOG.debug("set XAttr: {}.", f);
-
         checkPermission(f, RangerAccessType.WRITE);
-
-        // First, determine whether the length of the name and value exceeds the limit.
-        if (name.getBytes(METADATA_ENCODING).length + value.length > MAX_XATTR_SIZE) {
-            throw new HadoopIllegalArgumentException(String.format("The maximum combined size of " +
-                            "the name and value of an extended attribute in bytes should be less than or equal to %d",
-                    MAX_XATTR_SIZE));
-        }
-
-        Path absolutePath = makeAbsolute(f);
-
-        String key = pathToKey(absolutePath);
-        FileMetadata fileMetadata = store.retrieveMetadata(key);
-        if (null == fileMetadata) {
-            throw new FileNotFoundException("File or directory doesn't exist: " + f);
-        }
-        boolean xAttrExists = (null != fileMetadata.getUserAttributes()
-                && fileMetadata.getUserAttributes().containsKey(name));
-        XAttrSetFlag.validate(name, xAttrExists, flag);
-        if (fileMetadata.isFile()) {
-            store.storeFileAttribute(key, name, value);
-        } else {
-            store.storeDirAttribute(key, name, value);
-        }
+        this.actualImplFS.setXAttr(f, name, value, flag);
     }
 
     /**
@@ -1177,22 +476,8 @@ public class CosFileSystem extends FileSystem {
     @Override
     public byte[] getXAttr(Path f, String name) throws IOException {
         LOG.debug("get XAttr: {}.", f);
-
         checkPermission(f, RangerAccessType.READ);
-
-        Path absolutePath = makeAbsolute(f);
-
-        String key = pathToKey(absolutePath);
-        FileMetadata fileMetadata = store.retrieveMetadata(key);
-        if (null == fileMetadata) {
-            throw new FileNotFoundException("File or directory doesn't exist: " + f);
-        }
-
-        if (null != fileMetadata.getUserAttributes()) {
-            return fileMetadata.getUserAttributes().get(name);
-        }
-
-        return null;
+        return this.actualImplFS.getXAttr(f, name);
     }
 
     /**
@@ -1206,45 +491,15 @@ public class CosFileSystem extends FileSystem {
     @Override
     public Map<String, byte[]> getXAttrs(Path f, List<String> names) throws IOException {
         LOG.debug("get XAttrs: {}.", f);
-
         checkPermission(f, RangerAccessType.READ);
-
-        Path absolutePath = makeAbsolute(f);
-
-        String key = pathToKey(absolutePath);
-        FileMetadata fileMetadata = store.retrieveMetadata(key);
-        if (null == fileMetadata) {
-            throw new FileNotFoundException("File or directory doesn't exist: " + f);
-        }
-
-        Map<String, byte[]> attrs = null;
-        if (null != fileMetadata.getUserAttributes()) {
-            attrs = new HashMap<>();
-            for (String name : names) {
-                if (fileMetadata.getUserAttributes().containsKey(name)) {
-                    attrs.put(name, fileMetadata.getUserAttributes().get(name));
-                }
-            }
-        }
-
-        return attrs;
+        return this.actualImplFS.getXAttrs(f, names);
     }
 
     @Override
     public Map<String, byte[]> getXAttrs(Path f) throws IOException {
         LOG.debug("get XAttrs: {}.", f);
-
         checkPermission(f, RangerAccessType.READ);
-
-        Path absolutePath = makeAbsolute(f);
-
-        String key = pathToKey(absolutePath);
-        FileMetadata fileMetadata = store.retrieveMetadata(key);
-        if (null == fileMetadata) {
-            throw new FileNotFoundException("File or directory doesn't exist: " + f);
-        }
-
-        return fileMetadata.getUserAttributes();
+        return this.actualImplFS.getXAttrs(f);
     }
 
     /**
@@ -1257,45 +512,15 @@ public class CosFileSystem extends FileSystem {
     @Override
     public void removeXAttr(Path f, String name) throws IOException {
         LOG.debug("remove XAttr: {}.", f);
-
         checkPermission(f, RangerAccessType.WRITE);
-
-        Path absolutPath = makeAbsolute(f);
-
-        String key = pathToKey(absolutPath);
-        FileMetadata fileMetadata = store.retrieveMetadata(key);
-        if (null == fileMetadata) {
-            throw new FileNotFoundException("File or directory doesn't exist: " + f);
-        }
-
-        boolean xAttrExists = (null != fileMetadata.getUserAttributes()
-                && fileMetadata.getUserAttributes().containsKey(name));
-        if (xAttrExists) {
-            if (fileMetadata.isFile()) {
-                store.removeFileAttribute(key, name);
-            } else {
-                store.removeDirAttribute(key, name);
-            }
-        }
-
-        // Nothing to do if the specified attribute is not found.
+        this.actualImplFS.removeXAttr(f, name);
     }
 
     @Override
     public List<String> listXAttrs(Path f) throws IOException {
         LOG.debug("list XAttrs: {}.", f);
-
         checkPermission(f, RangerAccessType.READ);
-
-        Path absolutePath = makeAbsolute(f);
-
-        String key = pathToKey(absolutePath);
-        FileMetadata fileMetadata = store.retrieveMetadata(key);
-        if (null == fileMetadata) {
-            throw new FileNotFoundException("File or directory doesn't exist: " + f);
-        }
-
-        return new ArrayList<>(fileMetadata.getUserAttributes().keySet());
+        return this.actualImplFS.listXAttrs(f);
     }
 
     @Override
@@ -1332,7 +557,7 @@ public class CosFileSystem extends FileSystem {
         }
 
         Path absolutePath = makeAbsolute(f);
-        String allowKey = pathToKey(absolutePath);
+        String allowKey = CosHadoopFileSystem.pathToKey(absolutePath);
         if (allowKey.startsWith("/")) {
             allowKey = allowKey.substring(1);
         }
@@ -1351,20 +576,66 @@ public class CosFileSystem extends FileSystem {
         }
     }
 
-    @Override
-    public void close() throws IOException {
-        LOG.info("begin to close cos file system");
-        try {
-            super.close();
-        } finally {
-            this.store.close();
-            this.boundedIOThreadPool.shutdown();
-            this.boundedCopyThreadPool.shutdown();
-            BufferPool.getInstance().close();
+    public NativeFileSystemStore getStore() {
+        return this.store;
+    }
+
+    public boolean isPosixBucket() {
+        return this.isPosixBucket;
+    }
+
+    public boolean isCosProcess() {
+        return this.isCosProcess;
+    }
+
+    private Path makeAbsolute(Path path) {
+        if (path.isAbsolute()) {
+            return path;
+        }
+        return new Path(workingDir, path);
+    }
+
+    // exclude the ofs original config, filter the ofs config with COSN_CONFIG_TRANSFER_PREFIX
+    private void transferOfsConfig() {
+        // 1. list to get transfer prefix ofs config
+        Map<String, String> tmpConf = new HashMap<>();
+        for (Map.Entry<String, String> entry : this.getConf()) {
+            if (entry.getKey().startsWith(Constants.COSN_OFS_CONFIG_PREFIX)) {
+                this.getConf().unset(entry.getKey());
+            }
+            if (entry.getKey().startsWith(Constants.COSN_CONFIG_TRANSFER_PREFIX)) {
+                int pos = Constants.COSN_CONFIG_TRANSFER_PREFIX.length();
+                String subConfigKey = entry.getKey().substring(pos);
+                tmpConf.put(subConfigKey, entry.getValue());
+            }
+        }
+
+        // 2. trim the prefix and overwrite the config
+        for (Map.Entry<String, String> entry : tmpConf.entrySet()) {
+            LOG.info("transfer ofs config, key {}, value {}", entry.getKey(), entry.getValue());
+            this.getConf().set(entry.getKey(), entry.getValue());
         }
     }
 
-    public NativeFileSystemStore getStore() {
-        return this.store;
+    // some other override method support only by posix bucket
+    public void releaseFileLock(Path p) throws IOException {
+        if (!this.isCosProcess) {
+            if (this.actualImplFS instanceof CHDFSHadoopFileSystemAdapter) {
+                ((CHDFSHadoopFileSystemAdapter)this.actualImplFS).releaseFileLock(p);
+            } else {
+               throw new IOException("file system process but actual impl fs is not right");
+            }
+        } else {
+            throw new UnsupportedOperationException("Not supported currently");
+        }
+    }
+
+    @Override
+    public void close() throws IOException {
+        LOG.info("begin to close cos file system");
+        this.actualImplFS.close();
+        if (rangerQcloudObjectStorageStorageClient != null) {
+            rangerQcloudObjectStorageStorageClient.close();
+        }
     }
 }
