+++ src/main/java/org/apache/hadoop/fs/CosNConfigKeys.java
@@ -2,8 +2,8 @@ package org.apache.hadoop.fs;
 
 import org.apache.hadoop.classification.InterfaceAudience;
 import org.apache.hadoop.classification.InterfaceStability;
+import org.apache.hadoop.fs.cosn.Constants;
 import org.apache.hadoop.fs.cosn.Unit;
-
 
 /**
  * This class contains constants for configuration keys used in the cos file system.
@@ -12,7 +12,7 @@ import org.apache.hadoop.fs.cosn.Unit;
 @InterfaceStability.Unstable
 public class CosNConfigKeys extends CommonConfigurationKeys {
     public static final String USER_AGENT = "fs.cosn.user.agent";
-    public static final String DEFAULT_USER_AGENT = "cos-hadoop-plugin-v8.0.5";
+    public static final String DEFAULT_USER_AGENT = "cos-hadoop-plugin-v8.1.0";
 
     public static final String TENCENT_EMR_VERSION_KEY = "fs.emr.version";
 
@@ -99,10 +99,8 @@ public class CosNConfigKeys extends CommonConfigurationKeys {
     public static final int DEFAULT_MAX_CONNECTION_NUM = 2048;
 
     public static final String CUSTOMER_DOMAIN = "fs.cosn.customer.domain";
-    public static final String OPEN_CHECK_MERGE_BUCKET = "fs.cosn.check.merge.bucket";
-    public static final boolean DEFAULT_CHECK_MERGE_BUCKET = false;
-    public static final String MERGE_BUCKET_MAX_LIST_NUM = "fs.cosn.merge.bucket.max.list.num";
-    public static final int DEFAULT_MERGE_BUCKET_MAX_LIST_NUM = 5000;
+    public static final String POSIX_BUCKET_MAX_LIST_NUM = "fs.cosn.posix.bucket.max.list.num";
+    public static final int DEFAULT_POSIX_BUCKET_MAX_LIST_NUM = 5000;
     public static final String NORMAL_BUCKET_MAX_LIST_NUM = "fs.cosn.normal.bucket.max.list.num";
     public static final int DEFAULT_NORMAL_BUCKET_MAX_LIST_NUM = 999;
     public static final String COSN_SERVER_SIDE_ENCRYPTION_ALGORITHM = "fs.cosn.server-side-encryption.algorithm";
@@ -142,4 +140,11 @@ public class CosNConfigKeys extends CommonConfigurationKeys {
     // default disable emr v2 instance url.
     public static final String COSN_EMRV2_INSTANCE_PROVIDER_ENABLED = "fs.cosn.emrv2.instance.provider.enabled";
     public static final boolean DEFAULT_COSN_EMRV2_INSTANCE_PROVIDER_ENABLED = false;
+
+    public static final String COSN_POSIX_BUCKET_GATEWAY_MODE_ENABLED = "fs.cosn.posix.bucket.gateway.mode.enabled";
+    public static final boolean DEFAULT_COSN_POSIX_BUCKET_GATEWAY_MODE_ENABLED = false;
+    public static final String COSN_POSIX_BUCKET_NOGATEWAY_IMPL = "fs.cosn.posix.bucket.nogateway.impl";
+    public static final String DEFAULT_COSN_POSIX_BUCKET_NOGATEWAY_IMPL =
+            "com.qcloud.chdfs.fs.CHDFSHadoopFileSystemAdapter";
+
 }
