+++ src/main/java/org/apache/hadoop/fs/cosn/Constants.java
@@ -29,4 +29,7 @@ public final class Constants {
     // Environments variables for the COS secretId and secretKey.
     public static final String COSN_SECRET_ID_ENV = "COSN_SECRET_ID";
     public static final String COSN_SECRET_KEY_ENV = "COSN_SECRET_KEY";
+
+    public static final String COSN_OFS_CONFIG_PREFIX = "fs.ofs.";
+    public static final String COSN_CONFIG_TRANSFER_PREFIX = "fs.cosn.trsf.";
 }
